moves PetDetail and PetList to own files, out of Store
