+++ src/components/PetDetail.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const PetDetail = (props) => {
+  let petDetails = props.pets.map(x => {
+    return (
+      <article>
+        Pet type: {x.type} <br/>
+        Breed: {x.breed} <br/>
+        Cost: {x.cost} <br/>
+      </article>
+    );
+  });
+
+  return <section>{petDetails}</section>;
+};
+
+export default PetDetail;
